Avoid copying DC adminep RCA again on rehoming
On rehoming, the DC adminep RCA certificate is copied from the new
SystemController to the subcloud. This is done in:
- "Install DC admin endpoint root CA certificate", to the ca-trust
  directory ('/etc/pki/ca-trust/source/anchors'); and
- "Update DC admin endpoint root CA certificate in config permdir",
  to the staging directory ("/opt/platform/config/<sw_version>")

However, the task used to copy the files during bootstrap:
"Copy admin endpoint certificates to the shared filesystem directory"
ends up updating again the staging area with the contents from the
ca-trust folder.

This would be harmless, but a race condition with sysinv and puppet
can make puppet overwrite the ca-trust directory with invalid data,
which in turn is copied to the staging area by the task. When the
staging area is not overwritten with invalid data, the file in the
ca-trust directory is fixed by sysinv/puppet in the next manifest
executed.

Test plan:
PASS: Deploy DC SystemController + SX subcloud.
PASS: Rehome the subcloud.

Closes-bug: 2127875

Change-Id: Ie52389966ee178705080559c4b259af8a1af6590
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>