Fix auto-restore with remote install
Use the new mount_seed_iso extra variable in the restore playbook seed
ISO RVMC task conditional so it only attempts to mount the seed ISO
when restoring a pre-installed subcloud.
This fixes an issue where the seed ISO would also be mounted for
auto-restore with install and factory restore, where the subcloud is
reinstalled by the install playbook.

Test Plan:
01. PASS – Run standard subcloud restore. Verify it works as expected
    without calling the RVMC script to mount the seed ISO.
02. PASS – Run subcloud auto-restore on a pre-installed subcloud
    (without --with-install). Verify that the seed ISO is mounted by
    the RVMC script.
03. PASS – Run subcloud auto-restore with --with-install. Verify that
    the subcloud is remote installed by the install playbook and that
    the restore playbook does not mount the seed ISO.
04. PASS – Run subcloud factory restore. Verify that the subcloud is
    remote installed by the install playbook and that the restore
    playbook does not mount the seed ISO.

Closes-Bug: 2129547

Change-Id: I72701359de27561fa5082fab63386072828e981a
Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>