Fix metadata sync for subcloud remote install
The subcloud restore with remote install to N-1 release was not working
as expected due to the following 2 issues:

a) Bug introduced by commit:
https://review.opendev.org/c/starlingx/ansible-playbooks/+/945068
b) The software list command used to gather the list of releases and
their states does not work reliably right after the install. The USM API
may not be available by the time the task is executed. As a result, the
metadata sync flag is not set correctly.

This commit addresses
a) by reverting the previous change and leaving the original
implementation in place and ensuring the sync'ed metadata states are
correctly set for N-1 release
b) by using USM core python implementation to obtain the list of
releases

Test Plan:
PASS - Install a remote N-1 subcloud and verify that the software
       version and commit are synchronized correctly.

PASS - Install a remote N subcloud and verify that the software version
       and commit are synchronized correctly.

PASS - Restore a backup with install and verify that the software
       version and commit are synchronized correctly.

PASS - Install a remote N-1 subcloud when the system controller was
       deployed with N release and N-1 patches were uploaded

PASS - Install a remote N subcloud when N release patch has been applied
       on system controller

Note: Ensure to run the tests for both SX and DX subclouds.

Regression tests:
 - Install a remote N-1 subcloud when the system controller has been
   upgraded from N-1 + patch to N release.
 - Rehome a prepatched N-1 subcloud to N system controller and verify
   that it can be upgraded to N release

Closes-bug: 2129703

Change-Id: Ib73c04d94690440533e75779633e434c6ed82c30
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>