Use only the primary address for oidc_issuer_address
The change [1] introduced the oidc_issuer_address definition in the
bootstrap using a small snippet of  python that didn't consider the
dual-stack string format. This format contain a comma separating the
addresses for each protocol version (IPv4 or IPv6).

This change adjusts the code to consider that and select only the
primary address

[1] https://review.opendev.org/c/starlingx/ansible-playbooks/+/961555

Test Plan
=========
[PASS] bootstrap AIO-DX in single-stack
[PASS] bootstrap AIO-DX in dual-stack

Closes-Bug: 2131224

Change-Id: I7104d370c68bdcf06fa365cf27e90edbe6759500
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>