Change oslo policy_file to YAML
Changed policy_file references in service config files to yaml
files. The changes are fixing the documentation in the templates
to describe a file named 'policy.yaml'.

For the barbican config template, the puppet-barbican package will
set the file to '/etc/barbican/policy.yaml', so the value set was
removed. Neither file (nor .json or .yaml) is installed to be used,
so the change doesn't change functionality.

The usage of JSON policy file was deprecated in Victoria [1].

[1] https://docs.openstack.org/releasenotes/oslo.policy/victoria.html

Test plan:
PASS: Install AIO-SX.
      Checked that the values changed in the templates.

Story: 2011360
Task: 53164

Change-Id: Ifbdc73e4927b6381e51271f18cfd1a2d5d5c1ac6
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>