Add support for auto and factory restore without IPMI SEL event monitoring
This commit adds support to specify ipmi_sel_event_monitoring as false
in the restore values. When this option is false, both the auto and
factory restore will stop using IPMI SEL events to monitor the subcloud
state.

The auto-restore will wait for the full restore to complete, by waiting
for the subcloud to become reachable again and then checking for an
auto-restore complete flag.

The factory-restore will run remotely through the system controller,
allowing the system controller to monitor the full procedure. Since
the OAM IP can be different for the factory backup, the process will
also update the OAM IP from the factory default to the current value
so the system controller is able to communicate with the subcloud post
restore.

Test Plan:
01. PASS – Update the install values with ipmi_sel_event_monitoring as
    False and run auto-restore. Verify that the subcloud is restored
    and becomes online.
02. PASS - Update the install values with ipmi_sel_event_monitoring as
    False and run factory restore. Verify that the subcloud is restored
    to the factory default and becomes reachable by the system
    controller.
03. PASS - Remove ipmi_sel_event_monitoring from the restore values and
    run both factory and auto restore, verify that it defaults to using
    IPMI SEL events for monitoring and that both operations complete
    successfully.
04. PASS - Run the standard subcloud restore operation and verify it
    still works as expected.

Partial-Bug: 2134974

Change-Id: I8ddc41ec910eee075f761493f460f833a4db8d75
Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>