Fix undefined variable on mgmt reconfiguration
After a subcloud management network reconfiguration, the
system controller must also be updated to reach the subcloud using the
new management network. This is done with the command:
dcmanager subcloud update subcloud1 ...

In this scenario, the "mode" variable in the ansible playbook was not
set, causing the dcmanager update to fail. The fix adds a check to
ensure that the variable exists, since it is only required during
bootstrap.

Tests performed:
PASS AIO-SX fresh install
PASS AIO-SX subcloud management-network-reconfiguration

Closes-bug: 2135244
Change-Id: I19bebc9c8180e74f4f13c1e41d8c0e3e585409bd
Signed-off-by: Fabiano Correa Mercer <fabiano.correamercer@windriver.com>