Fix trident installation playbook
When a Fibre Channel (FC) backend is configured on Trident, the
behavior is identical to iSCSI, where the 'find_multipaths'
setting in /etc/multipath.conf needs to be 'no'. This was done
in [1], and needs to be done again when it is FC.

Therefore, whenever the storageDriverName is configured with
'ontap-san', which means it will be iscsi, nvme or fc, the
'run_multipath_check' flag will be set to 'true' and the
DaemonSet will be created.

In addition, tasks were also added to wait for the deletion
and creation of k8s resources, to avoid race conditions with
the trident CRDs.

[1]: https://review.opendev.org/c/starlingx/ansible-playbooks/+/967232

Test Plan:
- PASS: Configure 'ontap-nas' and verify that the
	daemonset has been created.

Closes-Bug: 2136765

Change-Id: I85ae4907b9c70fccbf19ea4f3593c982aa34946b
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>