Storage backend migration support playbooks
These playbooks allow the user to migrate from bare-metal ceph
to rook-ceph in three ways: redeploy, export-import or in-service
migration.

Also a playbook to support the migration in the subclouds is provided,
intended to be run from the system controller. All the other playbooks
are intended to be run in standalone systems.

Since there is a need for custom variables to be set for all hosts
in the standalone systems (controllers, workers and storage nodes),
the use of group_vars was used.

There is also the need for some custom ansible configurations
specific to this procedure, like fork = 1 and the use o ansible
cache. To fullfil this need, an environment variable specifing
the configuration file path must be set before tunning the playbooks.

With the custom configurations, a separate log file
"storage-backend-migration-ansible.log" was used, which will also
be collected by the "collect" command, as well as the cache.

Playbook command example:

ANSIBLE_CONFIG='/usr/share/ansible/stx-ansible/playbooks/vars/storage-backend-migration/ansible.cfg' \
ansible-playbook /usr/share/ansible/stx-ansible/playbooks/storage-backend-migration-redeploy.yaml

Test Plan:
  PASS: SX virtualized migrations - redeploy, export-import and in-service
  PASS: DX virtualized migrations - redeploy, export-import and in-service
  PASS: DX-PLUS virtualized migrations - redeploy, export-import and in-service
  PASS: STD virtualized migrations - redeploy, export-import and in-service
  PASS: STORAGE virtualized migration - redeploy
  PASS: STORAGE HW migration - redeploy
  PASS: DC HW migration - 3 SCs redeploy

Story: 2011640
Task: 53536

Change-Id: Ica60c8a578854ee972ae358be30b5889fabfb2af
Signed-off-by: Heitor Farhat <heitor.farhat@windriver.com>